Add background music
Add the Tetris theme music as background music. Make it possible to
disable it since it's most likely quite annoying.

It should be possible to use a compressed audio format but that
currently doesn't work with the version of ffmpeg on my development PC
so just use uncompressed audio. It's not huge after all.